# docs/mawaddah-openapi.json
openapi: 3.0.3
info:
  title: ''
  version: 0.0.0
paths:
...
  /api/appeals/my-appeals/:
    get:
      operationId: appeals_my_appeals_retrieve
      description: List appeals created by the current user.
      tags:
      - appeals
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
          description: ''
  /api/appeals/reviewable/:
    get:
      operationId: appeals_reviewable_retrieve
      description: List appeals visible to Shura for review (pending).
      tags:
      - appeals
      security:
      - cookieAuth: []
      - tokenAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appeal'
          description: ''
...
components:
  schemas:
    Appeal:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        title:
          type: string
          maxLength: 100
        description:
          type: string
          nullable: true
        category:
          $ref: '#/components/schemas/CategoryEnum'
        amount_requested:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        is_monthly:
          type: boolean
        months_required:
          type: integer
          nullable: true
        status:
          allOf:
          - $ref: '#/components/schemas/StatusEnum'
          readOnly: true
        is_urgent:
          type: boolean
          readOnly: true
        created_by:
          type: integer
          readOnly: true
        beneficiary:
          type: integer
          readOnly: true
        approved_by:
          type: integer
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - amount_requested
      - approved_by
      - beneficiary
      - category
      - created_at
      - created_by
      - id
      - is_urgent
      - status
      - title
      - updated_at
...
    CategoryEnum:
      enum:
      - house_rent
      - school_fee
      - medical
      - utility_bills
      - debt
      - business_support
      - death_support
      - other
      type: string
      description: |-
        * `house_rent` - House Rent
        * `school_fee` - School Fee
        * `medical` - Medical
        * `utility_bills` - Utility Bills
        * `debt` - Debt
        * `business_support` - Business Support
        * `death_support` - Death Support
        * `other` - Other
...
    StatusEnum:
      enum:
      - pending
      - approved
      - rejected
      - fulfilled
      - expired
      type: string
      description: |-
        * `pending` - Pending
        * `approved` - Approved
        * `rejected` - Rejected
        * `fulfilled` - Fulfilled
        * `expired` - Expired
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    tokenAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Token-based authentication with required prefix "Token"